data-drift/metrics_app/requirements.txt
dataclasses
Flask
pandas
Werkzeug
requests
prometheus_client
pyyaml
evidently